requirements.txt
boto3==1.17.51
botocore==1.20.51
click==7.1.2
jmespath==0.10.0
python-dateutil==2.8.1
python-decouple==3.4
s3transfer==0.3.7
six==1.15.0
urllib3==1.26.4
